# additional changes for optimization