--- a/rust/Cargo.toml
+++ b/rust/Cargo.toml
@@ -4,6 +4,8 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+ahash = "0.8.11"
+compact_str = "0.9.0"
 mimalloc = "0.1.46"
 num_cpus = "1.16.0"
 tokio = { version = "1.44.2", features = ["full"] }