Added some additional changes to the file: 

```ahash = "0.8.11"```
 is a more efficient hashing algorithm than the default one used in std::HashMap; because it is able to take advantage of AES instruction. [Read more here](https://github.com/tkaitchuck/aHash). Also includes benchmarks for additional hashers.

```compact_str = "0.9.0"```
a more efficient way to store strings directly on the stack instead of heap allocation. [Read more here](https://github.com/ParkMyCar/compact_str)

Original: 

```
- finished in 33.5618998s, task avg 1.209623ms, min 1.0107ms, max 62.1091ms
- finished in 31.4463629s, task avg 1.136508ms, min 1.0116ms, max 33.6175ms
- finished in 31.5273688s, task avg 1.139639ms, min 1.0101ms, max 29.6652ms
- finished in 33.5166868s, task avg 1.205907ms, min 1.0094ms, max 57.3404ms
```

After the few changes: 

```
- finished in 23.9794508s, task avg 890.473µs, min 633µs, max 56.4037ms
- finished in 22.0231254s, task avg 821.608µs, min 632.3µs, max 40.4787ms
- finished in 20.2079099s, task avg 757.14µs, min 634.9µs, max 44.6473ms
- finished in 23.5560827s, task avg 874.619µs, min 637.7µs, max 80.4039ms
```

For extra optimization I preallocated the **HashMap::with_capacity** of each item, I know you said you didn't ideally want this, but for the sake of showing benchmarks worth noting:

After capacity changes:
```
- finished in 13.4907308s, task avg 512.195µs, min 454.6µs, max 32.6919ms
- finished in 13.6798834s, task avg 518.547µs, min 454.7µs, max 32.5331ms
- finished in 13.2117127s, task avg 502.329µs, min 455.3µs, max 29.5416ms
- finished in 13.3526669s, task avg 507.706µs, min 455.1µs, max 59.3351ms
```