@curvednebula 2025-04-23
I've made corresponding changes in Go impl too.